personalization-system/requirements.txt
appdirs==1.4.4
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.6.3
blis==0.7.9
branca==0.6.0
bs4==0.0.1
cached-property==1.5.2
cachetools==5.2.1
catalogue==2.0.8
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
confection==0.0.4
cycler==0.11.0
cymem==2.0.7
datascience==0.17.5
db-dtypes==1.0.5
decorator==5.1.1
dm-tree==0.1.8
docutils==0.16
fastai==2.7.10
fastcore==1.5.27
fastdownload==0.0.7
fastjsonschema==2.16.2
fastprogress==1.0.3
Flask==1.1.4
folium==0.14.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.4.1
google-cloud-bigquery-storage==2.18.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
HeapDict==1.0.1
html5lib==1.0.1
httpimport==0.5.18
httplib2==0.17.4
idna==2.10
imageio==2.24.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
intel-openmp==2022.2.1
ipython==7.34.0
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==2.11.3
joblib==1.2.0
jsonschema==4.3.3
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-vis==0.4.1
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.2.2
matplotlib-inline==0.1.6
matplotlib-venn==0.11.7
mkl==2019.0
multidict==6.0.4
murmurhash==1.0.9
networkx==2.6.3
nltk==3.7
numpy==1.21.6
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
pandas-datareader==0.9.0
pandas-gbq==0.17.9
pandas-profiling==1.4.1
param==1.12.3
parso==0.8.3
pathlib==1.0.1
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.12.0
preshed==3.0.8
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.21.12
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydata-google-auth==1.5.0
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-louvain==0.16
python-slugify==7.0.0
python-utils==3.4.5
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.6.2
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools-git==1.2
six==1.16.0
sklearn-pandas==1.8.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
tenacity==8.2.2
text-unidecode==1.3
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
typer==0.7.0
urllib3==1.24.3
wasabi==0.10.1
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==1.0.1
wordcloud==1.8.2.2
xgboost==0.90
zipp==3.11.0
